LayoutTests/css3/filters/custom/custom-filter-crash-inline-computed-style.html

<!doctype html>
<html>
<head>
    <script src="../../../resources/js-test-pre.js"></script>
    <script>
        if (window.testRunner) {
            testRunner.overridePreference("WebKitCSSCustomFilterEnabled", "1");
            testRunner.overridePreference("WebKitWebGLEnabled", "1");
        }
    </script>
    <style>
        #filter-target {
            -webkit-filter: custom(url(no-shader-needed.vs), transform scale(2));
        }
    </style>
</head>
<body>
    <p id="description"></p>
    <div id="console"></div>
    <span id="filter-target"></span>
    <script>
        description("This test verifies that we don't assert when we generate the computed style\
            for an element that meets the following conditions:<br/>\
            (1) The element has a custom filter containing a transform parameter.<br/>\
            (2) The element does not have a CSS box. For example, the element is display: inline.");

        var element = document.getElementById("filter-target");
        shouldBeNonNull("element");

        var filterPropertyCSSValue = getComputedStyle(element).getPropertyCSSValue("-webkit-filter");
        testPassed("Computing the filter style for the element did not crash the browser.");
        shouldBeNonNull("filterPropertyCSSValue");
    </script>
    <script src="../../../resources/js-test-post.js"></script>
</body>
</html>
